+++ packages/rust/conundrum_db/src/vector/models/ecosystem_data/ecosystem_setting_key/settings_client.rs
@@ -45,7 +45,7 @@ impl SettingsClient {
         }
     }
 
-    pub async fn max_sync_threads(database: ArcMutexDB) -> DatabaseResult<usize> {
+    pub async fn max_sync_threads(database: ArcMutexDB) -> DatabaseResult<u16> {
         let setting = SettingsClient::get_setting(UniqueSettingKey::MaxSyncThreads, Arc::clone(&database)).await;
         match setting.data {
             Setting::Sync(k) => match k {
