+++ packages/rust/conundrum_db/src/vector/models/ecosystem_data/ecosystem_setting_key/sync_setting_key.rs
@@ -21,7 +21,7 @@ pub enum SyncSettingKey {
     AutoSyncOnNewMsg(bool),
     /// The maximum number of threads to be used while syncing. Note that this
     /// will default to the maximum number of threads available if set to 0.
-    MaxSyncThreads(usize),
+    MaxSyncThreads(u16),
 }
 
 impl EcosystemSettingKey for SyncSettingKey {
